+++ components/CPU.tsx
