+++ components/Memory.tsx
